move some methods to internal
