+++ internal/internal.go
@@ -1,4 +1,4 @@
-package utils
+package internal
 
 import (
 	"sort"
